+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Hidden Paths property.
+
+### Fixed
+
+- Skeleton display.
+
 ## [2.21.0] - 2026-07-02
 
 ### Added